+++ CHANGELOG.md
@@ -16,3 +16,7 @@
     - Creating UseCase.
     - Creating Adapter.
     - Creating Controller.
+- CRUD - Delete
+    - Creating UseCase.
+    - Creating Adapter.
+    - Creating Controller.